@@ -1,16 +1,27 @@
 {
   "name": "react-native-calendars",
   "version": "1.22.0",
-  "main": "src/index.js",
+  "main": "src/index.ts",
+  "types": "src/index.d.ts",
   "description": "React Native Calendar Components",
   "scripts": {
-    "test": "jasmine src/*.spec.js && npm run lint",
-    "lint": "eslint src/ example/src",
+    "build": "yarn build:tsc && yarn build:docs",
+    "build:ts": "tsc",
+    "build:dev": "tsc --noEmit",
+    "build:docs": "node ./scripts/build-docs.js",
+    "e2e": "node ./scripts/test-e2e.js --release",
+    "test": "yarn lint && yarn unit",
+    "test:watch": "jest --watch",
+    "test:perf": "TEST_RUNNER_ARGS='--testMatch **/*.perf.js' reassure measure",
+    "unit": "jest",
+    "lint": "eslint src/ example/src --ext .tsx,.ts,.js",
     "release": "node ./scripts/release.js",
     "start": "react-native start",
     "ios": "react-native run-ios",
     "android": "react-native run-android",
-    "xcode": "open ios/CalendarsExample.xcworkspace/"
+    "xcode": "open ios/CalendarsExample.xcworkspace/",
+    "clean": "rm yarn-lock && rm -rf ./node_modules && yarn",
+    "pod-install": "pod install --project-directory=ios"
   },
   "repository": {
     "type": "git",
@@ -22,19 +33,67 @@
   "author": "Wix.com",
   "license": "MIT",
   "dependencies": {
-    "lodash": "^4.0.0",
+    "hoist-non-react-statics": "^3.3.1",
+    "lodash": "^4.17.15",
+    "memoize-one": "^5.2.1",
     "prop-types": "^15.5.10",
+    "react-native-safe-area-context": "4.5.0",
+    "react-native-swipe-gestures": "^1.0.5",
+    "recyclerlistview": "^4.0.0",
     "xdate": "^0.8.0"
   },
   "devDependencies": {
-    "babel-eslint": "^7.2.3",
-    "eslint": "^6.5.1",
-    "eslint-plugin-react": "^7.0.0",
-    "jasmine": "^2.5.2",
-    "react": "16.9.0",
-    "react-native": "0.61.4",
-    "react-native-navigation": "^3.5.1",
+    "@babel/core": "^7.20.0",
+    "@babel/eslint-parser": "^7.13.4",
+    "@babel/preset-env": "^7.20.0",
+    "@babel/runtime": "^7.20.0",
+    "@react-native/babel-preset": "0.73.21",
+    "@react-native/eslint-config": "0.73.2",
+    "@react-native/metro-config": "0.73.5",
+    "@react-native/typescript-config": "0.73.1",
+    "@testing-library/react-native": "^11.0.0",
+    "@tsconfig/docusaurus": "^1.0.4",
+    "@types/jest": "^29.2.1",
+    "@types/lodash": "^4.14.170",
+    "@types/react": "^18.2.6",
+    "@types/react-native": "^0.64.19",
+    "@types/react-test-renderer": "^18.2.0",
+    "@types/xdate": "^0.8.32",
+    "@typescript-eslint/eslint-plugin": "^5.27.1",
+    "@typescript-eslint/parser": "^5.27.1",
+    "@welldone-software/why-did-you-render": "^6.0.3",
+    "babel-eslint": "^10.1.0",
+    "babel-jest": "^29.6.3",
+    "babel-plugin-module-resolver": "^5.0.2",
+    "detox": "^18.0.0",
+    "eslint": "8.19.0",
+    "eslint-config-prettier": "^6.11.0",
+    "eslint-plugin-jest": "^26.0.0",
+    "eslint-plugin-prettier": "^4.2.1",
+    "eslint-plugin-react": "^7.24.0",
+    "eslint-plugin-react-native": "^4.0.0",
+    "eslint-plugin-testing-library": "^7.1.1",
+    "husky": "^4.2.5",
+    "jest": "^29.6.3",
+    "jest-date-mock": "^1.0.8",
+    "lint-staged": "^10.2.11",
+    "metro-react-native-babel-preset": "0.73.7",
+    "mocha": "^7.1.0",
+    "prettier": "2.8.8",
+    "prettier-eslint": "^16.3.0",
+    "react": "18.2.0",
+    "react-native": "0.73.11",
+    "react-recipes": "^1.4.0",
+    "react-test-renderer": "18.2.0",
+    "reassure": "^0.4.1",
     "semver": "5.x.x",
-    "shell-utils": "1.x.x"
+    "shell-utils": "1.x.x",
+    "typescript": "5.0.4"
+  },
+  "optionalDependencies": {
+    "moment": "^2.29.4"
+  },
+  "engines": {
+    "node": ">=18"
   }
 }
